backend/triangle/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-templated-mail==1.1.1
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
Faker==25.4.0
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.0
prompt_toolkit==3.0.45
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==2.0.0
setuptools==70.0.0
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
